added "default_game" packaging in entities and refactored
