+++ src/main/java/entities/default_game/HardMaze.java
@@ -1,4 +1,4 @@
-package entities;
+package entities.default_game;
 
 /**
  * Hard level maze.
