+++ src/main/java/entities/default_game/IDrawable.java
@@ -1,4 +1,4 @@
-package entities;
+package entities.default_game;
 
 /**
  * An interface to draw entities on the game screen.
