+++ src/main/java/entities/default_game/MediumMaze.java
@@ -1,4 +1,4 @@
-package entities;
+package entities.default_game;
 
 /**
  * Medium level maze.
