+++ src/main/java/entities/default_game/Player.java
@@ -1,4 +1,4 @@
-package entities;
+package entities.default_game;
 
 /**
  * Represents the player. Data will be stored in another class.
